Implement loading state and post creation dialog
In the post list, conditional rendering was added to check if posts exist. In the event that it is empty, a message will be displayed to inform the user. In the App.vue, changes were made to create a dialog box for post creation, instead of the form being displayed outright. A loading state was introduced when fetching for posts. Axios was added to fetch posts from 'https://jsonplaceholder.typicode.com/posts?_limit=10'. This changes were made to improve user experience and interface interaction.